Move API error checking into API call function
This moves a lot of error checking out of the higher level functions

The API request now loops if bad status code, instead of old behavior which raised an error. However, there's a hard-coded 5 request limit. If that is reached, then it an error is thrown. Don't want to get stuck in a continuous loop of API requests