Refactor/rename MPL vis module
This again simplifies the OOP layout of this module

Refactor the MPL object into a bunch of functions with the first arg being the GHCND object instance. This means that there doesn't need to be a bunch of objects for different types of vis. However, there will potentially be many module files for every vis type