Reorganize raw data processing
Move some of the code from the df_proc method to update method

This way the df_proc method is only for numerical processing of the raw data, whereas the DataFrame creation and concat parts of the code are in the update method. This should make overriding the processing more straightforward in the future.